+++ client_mod/mods-unpacked/RampagingHippy-Archipelago/ap/brotato_ap_client.gd
@@ -19,6 +19,7 @@ var ApUpgradesProgress = load("res://mods-unpacked/RampagingHippy-Archipelago/pr
 var ApLootCrateProgress = load("res://mods-unpacked/RampagingHippy-Archipelago/progress/loot_crates.gd")
 var ApWavesProgress = load("res://mods-unpacked/RampagingHippy-Archipelago/progress/waves.gd")
 var ApWinsProgress = load("res://mods-unpacked/RampagingHippy-Archipelago/progress/wins.gd")
+var ApSavedRunsProgress = load("res://mods-unpacked/RampagingHippy-Archipelago/progress/saved_runs.gd")
 var GodotApClientDebugSettings = load("res://mods-unpacked/RampagingHippy-Archipelago/ap/debug.gd")
 
 var game_state
@@ -37,6 +38,10 @@ var common_loot_crate_progress
 var legendary_loot_crate_progress
 var waves_progress
 var wins_progress
+var saved_runs_progress
+
+# Holder for all the trackers together so we can perform operations over all of them easily.
+var _progress_trackers: Array
 
 signal on_connection_refused(reasons)
 
@@ -55,9 +60,42 @@ func _init(websocket_client, config).(websocket_client, config):
 	legendary_loot_crate_progress = ApLootCrateProgress.new(self, game_state, "legendary")
 	waves_progress = ApWavesProgress.new(self, game_state)
 	wins_progress = ApWinsProgress.new(self, game_state)
+	saved_runs_progress = ApSavedRunsProgress.new(self, game_state)
 	debug = GodotApClientDebugSettings.new()
 
+	_progress_trackers = [
+		character_progress,
+		shop_slots_progress,
+		shop_lock_buttons_progress,
+		enemy_xp_progress,
+		gold_progress,
+		xp_progress,
+		items_progress,
+		upgrades_progress,
+		common_loot_crate_progress,
+		legendary_loot_crate_progress,
+		waves_progress,
+		wins_progress,
+		saved_runs_progress,
+	]
+
 	ModLoaderLog.debug("Brotato AP adapter initialized", _LOG_NAME)
+
+
+func export_run_specific_progress_data() -> Dictionary:
+	## Gather all run-specific data from the progress trackers into a single dictionary.
+	##
+	## This is meant to complement Brotato's built-in save data, so we can properly load
+	## the save and restore any progress data that's specific to that run.
+	var data = {}
+	for progress in _progress_trackers:
+		var run_data = progress.export_run_specific_progress_data()
+		data.merge(run_data)
+	return data
+
+func load_run_specific_progress_data(data: Dictionary):
+	for progress in _progress_trackers:
+		progress.load_run_specific_progress_data(data)
 
 func connected_to_multiworld() -> bool:
 	# Convenience method to check if connected to AP, so other scenes don't need to