+++ client_mod/mods-unpacked/RampagingHippy-Archipelago/progress/items.gd
@@ -22,6 +22,8 @@
 extends "res://mods-unpacked/RampagingHippy-Archipelago/progress/_base.gd"
 class_name ApItemsProgress
 
+const SAVE_DATA_KEY = "progress_items"
+
 signal item_received(item_tier)
 
 var received_items_by_tier: Dictionary = {
@@ -50,11 +52,10 @@ func process_ap_item(item_tier: int, player_index: int) -> int:
 	## The returned value won't necessarily match the current wave. The "<rarity> Item"
 	## items in the multiworld are meant to be evenly distributed over 20 waves, which
 	## is what this function handles.
-	
 	# Clamp the index access in case the player added more items with the admin console
 	# or something.
 	var wave_for_next_item = min(
-		processed_items_by_player_by_tier[player_index][item_tier], 
+		processed_items_by_player_by_tier[player_index][item_tier],
 		wave_per_game_item[item_tier].size()
 	)
 	processed_items_by_player_by_tier[player_index][item_tier] += 1
@@ -102,3 +103,25 @@ func on_run_started(character_ids: Array):
 				Tier.LEGENDARY: 0
 			}
 		)
+
+func export_run_specific_progress_data() -> Dictionary:
+	var data = {
+		SAVE_DATA_KEY: {
+			"processed_items_by_player_by_tier": processed_items_by_player_by_tier.duplicate()
+		}
+	}
+	return data
+
+func load_run_specific_progress_data(data: Dictionary):
+	processed_items_by_player_by_tier = []
+	var saved_processed_items = data[SAVE_DATA_KEY]["processed_items_by_player_by_tier"]
+	# If the data was loaded from JSON, the keys will be strings instead of ints
+	var is_string_key:bool  = not saved_processed_items[0].has(Tier.COMMON)
+	var tier_keys = [Tier.COMMON, Tier.UNCOMMON, Tier.RARE, Tier.LEGENDARY]
+	for entry in saved_processed_items:
+		var player_processed_items = {}
+		for tk in tier_keys:
+			var entry_key = str(tk) if is_string_key else tk
+			var tier_value = entry[entry_key]
+			player_processed_items[tk] = tier_value
+		processed_items_by_player_by_tier.push_back(player_processed_items)